+++ metadata.php
@@ -13,6 +13,7 @@ use OxidSolutionCatalysts\Adyen\Controller\Admin\AdminOrderController;
 use OxidSolutionCatalysts\Adyen\Controller\OrderController;
 use OxidSolutionCatalysts\Adyen\Controller\PaymentController;
 use OxidSolutionCatalysts\Adyen\Core\Module;
+use OxidSolutionCatalysts\Adyen\Core\Email;
 use OxidSolutionCatalysts\Adyen\Core\ViewConfig;
 use OxidSolutionCatalysts\Adyen\Model\Address;
 use OxidSolutionCatalysts\Adyen\Model\Country;
@@ -55,6 +56,7 @@ $aModule = [
         \OxidEsales\Eshop\Application\Model\User::class => User::class,
         // core
         \OxidEsales\Eshop\Core\ViewConfig::class => ViewConfig::class,
+        \OxidEsales\Eshop\Core\Email::class => Email::class,
         // admin-controller
         \OxidEsales\Eshop\Application\Controller\Admin\OrderList::class => OrderList::class,
         \OxidEsales\Eshop\Application\Controller\Admin\OrderArticle::class => OrderArticle::class,
@@ -74,6 +76,20 @@ $aModule = [
         'AdyenWebhookController' => AdyenWebhookController::class
     ],
     'settings' => [
+        [
+            'group'       => 'osc_adyen_mails',
+            'name'        => ModuleSettings::REFUND_MAIL_RECIPIENT,
+            'type'        => 'select',
+            'constraints' => '0|1|2|3',
+            'value'       => '0'
+        ],
+        [
+            'group'       => 'osc_adyen_mails',
+            'name'        => ModuleSettings::CANCEL_MAIL_RECIPIENT,
+            'type'        => 'select',
+            'constraints' => '0|1|2|3',
+            'value'       => '0'
+        ],
         [
             'group'       => 'osc_adyen_API',
             'name'        => ModuleSettings::OPERATION_MODE,